+++ src/AaveV3Arbitrum.sol
@@ -415,9 +415,9 @@ library AaveV3ArbitrumAssets {
 }
 
 library AaveV3ArbitrumEModes {
-  uint256 internal constant NONE = 0;
+  uint8 internal constant NONE = 0;
 
-  uint256 internal constant STABLECOINS = 1;
+  uint8 internal constant STABLECOINS = 1;
 
-  uint256 internal constant ETH_CORRELATED = 2;
+  uint8 internal constant ETH_CORRELATED = 2;
 }
