+++ src/AaveV3Harmony.sol
@@ -243,7 +243,7 @@ library AaveV3HarmonyAssets {
 }
 
 library AaveV3HarmonyEModes {
-  uint256 internal constant NONE = 0;
+  uint8 internal constant NONE = 0;
 
-  uint256 internal constant STABLECOINS = 1;
+  uint8 internal constant STABLECOINS = 1;
 }
