+++ src/AaveV3Polygon.sol
@@ -516,11 +516,11 @@ library AaveV3PolygonAssets {
 }
 
 library AaveV3PolygonEModes {
-  uint256 internal constant NONE = 0;
+  uint8 internal constant NONE = 0;
 
-  uint256 internal constant STABLECOINS = 1;
+  uint8 internal constant STABLECOINS = 1;
 
-  uint256 internal constant MATIC_CORRELATED = 2;
+  uint8 internal constant MATIC_CORRELATED = 2;
 
-  uint256 internal constant ETH_CORRELATED = 3;
+  uint8 internal constant ETH_CORRELATED = 3;
 }
