+++ src/AaveV3Sepolia.sol
@@ -241,7 +241,7 @@ library AaveV3SepoliaAssets {
 }
 
 library AaveV3SepoliaEModes {
-  uint256 internal constant NONE = 0;
+  uint8 internal constant NONE = 0;
 
-  uint256 internal constant STABLECOINS = 1;
+  uint8 internal constant STABLECOINS = 1;
 }
